fix: move settings from left-side bar to settings page
