+++ src/views/Settings/styles.module.css
@@ -13,13 +13,48 @@
             .container {
                 display: flex;
                 flex-direction: column;
-                background-color: var(--color-quaternary);
-                padding: var(--spacing-xs);
-                gap: var(--spacing-xs);
+                gap: var(--width-separator-lg);
 
                 .work-item {
                     background-color: var(--color-foreground);
                 }
+            }
+        }
+    }
+
+
+    .attribute-list {
+        display: flex;
+        flex-direction: column;
+        gap: var(--width-separator-lg);
+
+        &.dragging-over {
+            outline: var(--width-separator-md) solid var(--color-separator);
+        }
+
+        .attribute {
+            display: flex;
+            align-items: center;
+            background-color: var(--color-foreground);
+            padding: var(--spacing-xs) var(--spacing-sm);
+            gap: var(--spacing-xs);
+
+            &.dragging {
+                opacity: 0.8;
+                z-index: 1;
+                box-shadow: var(--box-shadow-md);
+            }
+
+            .drag-handle {
+                cursor: grab;
+            }
+
+            &:hover {
+                background-color: var(--color-tertiary);
+            }
+
+            .label {
+                flex-grow: 1;
             }
         }
     }
